Address review feedback on addon args and test naming
Rename list/string addon RPC argument types and fold PipelineOptions
tests into pipeline_test.go for clearer, shared helpers.
