+++ server/forge/addon/args.go
@@ -44,34 +44,16 @@ type argumentsNetrc struct {
 	R *modelRepo `json:"r"`
 }
 
-type argumentsActivateDeactivate struct {
-	U    *modelUser `json:"u"`
-	R    *modelRepo `json:"r"`
-	Link string     `json:"link"`
-}
-
-type argumentsBranchesPullRequests struct {
+type argumentsListOptions struct {
 	U *modelUser         `json:"u"`
 	R *modelRepo         `json:"r"`
 	P *model.ListOptions `json:"p"`
 }
 
-type argumentsBranchHead struct {
-	U      *modelUser `json:"u"`
-	R      *modelRepo `json:"r"`
-	Branch string     `json:"branch"`
-}
-
-type argumentsTagHead struct {
-	U   *modelUser `json:"u"`
-	R   *modelRepo `json:"r"`
-	Tag string     `json:"tag"`
-}
-
-type argumentsCommit struct {
-	U   *modelUser `json:"u"`
-	R   *modelRepo `json:"r"`
-	SHA string     `json:"sha"`
+type argumentsString struct {
+	U *modelUser `json:"u"`
+	R *modelRepo `json:"r"`
+	S string     `json:"s"`
 }
 
 type argumentsOrgMembershipOrg struct {
